Move watch item run logic from home to item row
